Extract `applyBandwidthLimitation` to a separate `bandwidth.ts` file
